refactor: Import via index helper
