+++ src/components/sign-up/screen/index.tsx
@@ -1,0 +1,5 @@
+export * from './screen-body';
+export * from './screen-actions';
+export * from './screen-content';
+export * from './screen-footer';
+export * from './screen-header';
